audio_editor/requirements.txt
amqp==5.0.2
asgiref==3.3.1
attrs==20.3.0
bcrypt==3.2.0
billiard==3.6.3.0
cached-property==1.5.2
celery==5.0.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
distro==1.5.0
Django==3.1.3
django-braces==1.14.0
django-cleanup==5.1.0
django-extensions==3.1.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docker==4.4.1
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==5.0.2
MarkupSafe==1.1.1
mutagen==1.45.1
numpy==1.19.5
oauthlib==3.1.0
packaging==20.7
paramiko==2.7.2
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
pydub==0.24.1
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sox==1.4.1
sqlparse==0.4.1
swagger-spec-validator==2.7.3
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
